+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/dccticketing/core/DccTicketingCoreModule.kt
@@ -4,18 +4,10 @@ import dagger.Module
 import dagger.Provides
 import dagger.Reusable
 import de.rki.coronawarnapp.dccticketing.core.server.DccTicketingApiV1
-import de.rki.coronawarnapp.http.HttpErrorParser
-import de.rki.coronawarnapp.http.config.HTTPVariables
-import de.rki.coronawarnapp.http.interceptor.RetryInterceptor
-import de.rki.coronawarnapp.http.interceptor.WebSecurityVerificationInterceptor
-import de.rki.coronawarnapp.risk.TimeVariables
-import okhttp3.Interceptor
+import de.rki.coronawarnapp.http.build
 import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import timber.log.Timber
-import java.util.concurrent.TimeUnit
 import javax.inject.Qualifier
 
 @Module
@@ -23,25 +15,7 @@ class DccTicketingCoreModule {
 
     @DccTicketingHttpClient
     @Provides
-    fun provideHttpClient(): OkHttpClient {
-        val interceptors: List<Interceptor> = listOf(
-            WebSecurityVerificationInterceptor(),
-            HttpLoggingInterceptor { message -> Timber.tag("OkHttp").v(message) }.apply {
-                setLevel(HttpLoggingInterceptor.Level.BODY)
-            },
-            RetryInterceptor(),
-            HttpErrorParser()
-        )
-
-        return OkHttpClient.Builder().apply {
-            connectTimeout(HTTPVariables.getHTTPConnectionTimeout(), TimeUnit.MILLISECONDS)
-            readTimeout(HTTPVariables.getHTTPReadTimeout(), TimeUnit.MILLISECONDS)
-            writeTimeout(HTTPVariables.getHTTPWriteTimeout(), TimeUnit.MILLISECONDS)
-            callTimeout(TimeVariables.getTransactionTimeout(), TimeUnit.MILLISECONDS)
-
-            interceptors.forEach { addInterceptor(it) }
-        }.build()
-    }
+    fun provideHttpClient(): OkHttpClient = OkHttpClient.Builder().build(enableLogging = true)
 
     @Reusable
     @Provides
