+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/http/HttpModule.kt
@@ -4,20 +4,12 @@ import dagger.Module
 import dagger.Provides
 import dagger.Reusable
 import de.rki.coronawarnapp.BuildConfig
-import de.rki.coronawarnapp.http.config.HTTPVariables
-import de.rki.coronawarnapp.http.interceptor.RetryInterceptor
-import de.rki.coronawarnapp.http.interceptor.WebSecurityVerificationInterceptor
-import de.rki.coronawarnapp.risk.TimeVariables
 import okhttp3.CipherSuite
 import okhttp3.ConnectionSpec
-import okhttp3.Interceptor
 import okhttp3.OkHttpClient
 import okhttp3.TlsVersion
-import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.converter.protobuf.ProtoConverterFactory
-import timber.log.Timber
-import java.util.concurrent.TimeUnit
 
 @Module
 class HttpModule {
@@ -25,25 +17,7 @@ class HttpModule {
     @Reusable
     @HttpClientDefault
     @Provides
-    fun defaultHttpClient(): OkHttpClient {
-        val interceptors: List<Interceptor> = listOf(
-            WebSecurityVerificationInterceptor(),
-            HttpLoggingInterceptor { message -> Timber.tag("OkHttp").v(message) }.apply {
-                if (BuildConfig.DEBUG) setLevel(HttpLoggingInterceptor.Level.BODY)
-            },
-            RetryInterceptor(),
-            HttpErrorParser()
-        )
-
-        return OkHttpClient.Builder().apply {
-            connectTimeout(HTTPVariables.getHTTPConnectionTimeout(), TimeUnit.MILLISECONDS)
-            readTimeout(HTTPVariables.getHTTPReadTimeout(), TimeUnit.MILLISECONDS)
-            writeTimeout(HTTPVariables.getHTTPWriteTimeout(), TimeUnit.MILLISECONDS)
-            callTimeout(TimeVariables.getTransactionTimeout(), TimeUnit.MILLISECONDS)
-
-            interceptors.forEach { addInterceptor(it) }
-        }.build()
-    }
+    fun defaultHttpClient(): OkHttpClient = OkHttpClient.Builder().build(BuildConfig.DEBUG)
 
     @Reusable
     @Provides
